memory-agent/requirements.txt
zep-cloud==3.2.0
python-dotenv==1.0.1
anthropic==0.37.1
openai